requirements.txt
Django==1.11.4
django-bower==5.2.0
django-filter==1.0.4
django-markdownx==2.0.21
django-ordered-model==1.4.2
django-widget-tweaks==1.4.1
Markdown==2.6.9
olefile==0.44
Pillow==4.2.1
pytz==2017.2
six==1.10.0
